Added Room Service

+++ src/MedX.Domain/Entities/Rooms/Room.cs
@@ -3,7 +3,7 @@ namespace MedX.Domain.Enitities;
 
 public class Room : Auditable
 {
-    public long RoomNumber { get; set; }
+    public int RoomNumber { get; set; }
     public int Quantity { get; set; }
 
     public ICollection<Patient> Patients { get; set; }
