+++ src/MedX.Service/DTOs/Rooms/RoomCreationDto.cs
@@ -2,6 +2,6 @@
 
 public class RoomCreationDto
 {
-    public long RoomNumber { get; set; }
+    public int RoomNumber { get; set; }
     public int Quantity { get; set; }
 }
