+++ src/MedX.Service/DTOs/Rooms/RoomUpdateDto.cs
@@ -3,6 +3,6 @@
 public class RoomUpdateDto
 {
     public long Id { get; set; }
-    public long RoomNumber { get; set; }
+    public int RoomNumber { get; set; }
     public int Quantity { get; set; }
 }
